# Improve pool.Endpoint.Equals() peformance