--- a/registry/registry_benchmark_test.go
+++ b/registry/registry_benchmark_test.go
@@ -69,6 +69,7 @@ func BenchmarkRegisterWith100KRoutes(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		r.Register("foo50000.example.com", fooEndpoint)
 	}
+	b.ReportAllocs()
 }
 
 func BenchmarkRegisterWithOneRoute(b *testing.B) {
@@ -81,4 +82,5 @@ func BenchmarkRegisterWithOneRoute(b *testing.B) {
 	for i := 0; i < b.N; i++ {
 		r.Register("foo.example.com", fooEndpoint)
 	}
+	b.ReportAllocs()
 }