<!-- Thanks for contributing to 'gorouter'. To speed up the process of reviewing your pull request please provide us with: -->

* A short explanation of the proposed change:

The change improves the `Endpoint.Equal()` function to use less time and no allocations during comparison. In profiler runs, @domdom82 identified that a lot of time is spent on the two `fmt.Sprintf` calls for comparison of the `tags` maps.

* An explanation of the use cases your change solves

Route registration messages that happen on route refresh are compared with the existing endpoints in the route registry for every received `router.register` message. This happens for every route by default every 20 seconds. In our scenarios we have up to 70k routes (on each gorouter) and such comparisons add up.

* Instructions to functionally test the behavior change using operator interfaces (BOSH manifest, logs, curl, and metrics)

Functionality is not changed. A benchmark has been added to show the changes.

* Expected result after the change

Route registration (and re-registration) should be a little faster. This will have measurable effect for large numbers of routes and route registration messages only.

* Current result before the change

Route registration takes about 8% of the processing time in Gorouter in our profiling examples. See https://github.com/cloudfoundry/routing-release/issues/366.

* Links to any other associated PRs


* [x] I have viewed signed and have submitted the Contributor License Agreement

* [x] I have made this pull request to the `main` branch

* [x] I have [run all the unit tests](https://github.com/cloudfoundry/routing-release#running-unit-and-integration-tests).

* [ ] (Optional) I have run Routing Acceptance Tests and Routing Smoke Tests

* [ ] (Optional) I have run CF Acceptance Tests