@peanball 2023-12-01
Some benchmark results:

There is a new benchmark, `pool_bench_test.go`:
```
BenchmarkEndpointEquals-16           	  695605	      1597 ns/op	     480 B/op	      16 allocs/op <- fmt.Sprintf
BenchmarkEndpointEquals-16           	13326138	        83.46 ns/op	       0 B/op	       0 allocs/op <- maps.Equal
```

While this is a microbenchmark and the difference between those two ops looks very stark, the pool endpoint insertion is a small part of the overall processing. Eliminating allocations should be a significant benefit however for this often-called function in terms of cache coherency, reducing memory access, etc.

Finally, the results of `registry_benchmark_test.go`:
```
old version
BenchmarkRegisterWith100KRoutes-16    	  939470	      1712 ns/op	    1546 B/op	      11 allocs/op
BenchmarkRegisterWithOneRoute-16      	  977769	      1208 ns/op	    1546 B/op	      11 allocs/op

new version
BenchmarkRegisterWith100KRoutes-16    	 1000000	      1312 ns/op	    1440 B/op	       7 allocs/op
BenchmarkRegisterWithOneRoute-16      	 1281951	       839.3 ns/op	    1440 B/op	       7 allocs/op
```

There are fewer allocations and generally less time per iteration with the new comparator.

The semantics of `Endpoint.Equals()` remain identical.