@maxmoehl 2023-12-03
When merging this into routing-release, please remember to bump [the `go` directive](https://github.com/cloudfoundry/routing-release/blob/69dba350cbc5d2a92ae3d5b15309d3b6899c0814/src/code.cloudfoundry.org/go.mod#L3) to at least `1.21.0`.