@ameowlia 2023-12-04
Oh no! This is failing go vet...

```
go vet ./...
# code.cloudfoundry.org/gorouter/route_test
gorouter/route/pool_bench_test.go:35:14: variable declaration copies lock value to endpoint3: code.cloudfoundry.org/gorouter/route.Endpoint contains sync.RWMutex
```